refactor: 更新ロジックをカスタムフックに移動 #37
